[twelf] make rules more like dependent lcf rules [notes below]
Refinement rules have many possible operational interpretations. One is
through Dependent LCF, which does not use unification. Another way to
interpret refinement rules is using logic programming and unification,
which is what the Twelf model does.

Because we intend the refinement rules to be compatible with the former
approach, we cannot "pattern match" on the extract in a premise (i.e.
generate unification constraints); instead, we choose to do all the
unification within the refinement logic by hand. This makes our rules
compatible with an interpretation not based on unification.